+++ applications/fishing-map/src/features/workspace/shared/LayerPanelContainer.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
+import HighlightPanel from '../highlight-panel/HighlightPanel'
+import styles from './LayerPanel.module.css'
+
+type LayerPanelProps = {
+  dataview: UrlDataviewInstance
+  children: React.ReactElement
+}
+
+function LayerPanelContainer({ dataview, children }: LayerPanelProps): React.ReactElement {
+  return (
+    <div className={styles.LayerPanelContainer}>
+      {children}
+      <HighlightPanel dataviewInstanceId={dataview.id} />
+    </div>
+  )
+}
+
+export default LayerPanelContainer