+++ applications/vessel-history/src/features/profile/components/activity/ActivityItem.tsx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react'
 import cx from 'classnames'
 import { Icon, IconButton } from '@globalfishingwatch/ui-components'
-import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
+import { RenderedEvent } from 'features/vessels/activity/vessels-activity.selectors'
 import { getEncounterStatus } from 'features/vessels/activity/vessels-activity.utils'
 import ActivityDate from './ActivityDate'
 import styles from './Activity.module.css'
@@ -14,16 +14,19 @@ interface EventProps {
 
 const ActivityItem: React.FC<EventProps> = ({
   event,
-  onInfoClick = () => { },
-  onMapClick = () => { },
+  onInfoClick = () => {},
+  onMapClick = () => {},
 }): React.ReactElement => {
   return (
     <Fragment>
       <div className={styles.event}>
-        <div className={cx(styles.eventIcon, styles[event.type], styles[getEncounterStatus(event)])}>
+        <div
+          className={cx(styles.eventIcon, styles[event.type], styles[getEncounterStatus(event)])}
+        >
           {event.type === 'encounter' && <Icon icon="event-encounter" type="default" />}
           {event.type === 'loitering' && <Icon icon="event-loitering" type="default" />}
           {event.type === 'fishing' && <Icon icon="event-fishing" type="default" />}
+          {event.type === 'port_visit' && <Icon icon="event-port-visit" type="default" />}
         </div>
         <div className={styles.eventData}>
           <ActivityDate event={event} />