+++ packages/dataviews-client/src/index.ts
@@ -3,12 +3,11 @@ export {
   default as resolveDataviews,
   resolveDataviewDatasetResource,
   resolveDataviewDatasetResources,
-  resolveDataviewEventsResources,
-  UrlDataviewInstance,
   mergeWorkspaceUrlDataviewInstances,
   resolveResourcesFromDatasetConfigs,
   getDataviewsForResourceQuerying,
-  TrackDatasetConfigs,
+  UrlDataviewInstance,
+  DatasetConfigsTransforms,
 } from './resolve-dataviews'
 export {
   getGeneratorConfig,